Task #758 - separate send & receive views, refactoring
