+++ ui/CMakeLists.txt
@@ -95,6 +95,18 @@ set(UI_SRC
     viewmodel/statusbar_view.cpp
     viewmodel/theme.h
     viewmodel/theme.cpp
+    viewmodel/receive_view.h
+    viewmodel/receive_view.cpp
+    viewmodel/send_view.h
+    viewmodel/send_view.cpp
+    viewmodel/status_holder.cpp
+    viewmodel/status_holder.h
+    viewmodel/tx_object.h
+    viewmodel/tx_object.cpp
+    viewmodel/payment_item.h
+    viewmodel/payment_item.cpp
+    viewmodel/qml_globals.h
+    viewmodel/qml_globals.cpp
 
     model/wallet_model.h
     model/wallet_model.cpp
