+++ ui/view/qml.qrc
@@ -36,6 +36,11 @@
         <file>settings.qml</file>
         <file>addresses.qml</file>
         <file>utxo.qml</file>
+        <file>receive.qml</file>
+        <file>receive_addrfail.qml</file>
+        <file>send.qml</file>
+        <file>send_confirm.qml</file>
+        <file>send_expired.qml</file>
         <file>assets/fonts/SF-Pro-Display-Black.otf</file>
         <file>assets/fonts/SF-Pro-Display-Light.otf</file>
         <file>assets/fonts/SF-Pro-Display-Semibold.otf</file>
