+++ ui/viewmodel/address_book_view.h
@@ -93,9 +93,7 @@ class AddressBookViewModel : public QObject
     Q_PROPERTY(QString contactSortRole READ contactSortRole WRITE setContactSortRole)
 
 public:
-
 	Q_INVOKABLE void deleteAddress(const QString& addr);
-    Q_INVOKABLE void copyToClipboard(const QString& text);
     Q_INVOKABLE void saveChanges(const QString& addr, const QString& name, bool isNever, bool makeActive, bool makeExpired);
     Q_INVOKABLE static QString generateQR(const QString& addr, uint width, uint height);
     Q_INVOKABLE static QString getLocaleName();
