+++ ui/viewmodel/qml_globals.cpp
@@ -1,0 +1,31 @@
+// Copyright 2019 The Beam Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+#include "qml_globals.h"
+#include <QMessageBox>
+#include <QApplication>
+
+QMLGlobals::QMLGlobals(QQmlEngine& engine)
+    : _engine(engine)
+{
+}
+
+void QMLGlobals::showMessage(QString message)
+{
+    QMessageBox::information(nullptr, "BeamWalletUI", message);
+}
+
+void QMLGlobals::copyToClipboard(const QString& text)
+{
+    QApplication::clipboard()->setText(text);
+}
