+++ ui/viewmodel/qml_globals.h
@@ -1,0 +1,30 @@
+// Copyright 2019 The Beam Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+#pragma once
+
+#include <QObject>
+#include <QQmlApplicationEngine>
+
+class QMLGlobals : public QObject
+{
+    Q_OBJECT
+public:
+    QMLGlobals(QQmlEngine&);
+
+    Q_INVOKABLE void showMessage(QString message);
+    Q_INVOKABLE void copyToClipboard(const QString& text);
+
+private:
+    QQmlEngine& _engine;
+};
